refactoring of bruteforce and password
