+++ src/screens/enter-password/enter-password.tsx
@@ -1,7 +1,6 @@
 import { Formik } from 'formik';
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback } from 'react';
 import { Text, View } from 'react-native';
-import { useDispatch } from 'react-redux';
 
 import { useBiometryAvailability } from '../../biometry/use-biometry-availability.hook';
 import { ButtonLargePrimary } from '../../components/button/button-large/button-large-primary/button-large-primary';
@@ -18,9 +17,9 @@ import { HIDE_SPLASH_SCREEN_TIMEOUT } from '../../config/animation';
 import { MaxPasswordAttemtps } from '../../config/system';
 import { FormPasswordInput } from '../../form/form-password-input';
 import { useDelayedEffect } from '../../hooks/use-delayed-effect.hook';
+import { usePasswordLock } from '../../hooks/use-password-lock.hook';
 import { useResetDataHandler } from '../../hooks/use-reset-data-handler.hook';
 import { useAppLock } from '../../shelter/use-app-lock.hook';
-import { setPasswordTimelock } from '../../store/security/security-actions';
 import { usePasswordAttempt, usePasswordTimelock } from '../../store/security/security-selectors';
 import { useBiometricsEnabledSelector } from '../../store/settings/settings-selectors';
 import { formatSize } from '../../styles/format-size';
@@ -33,32 +32,16 @@ import {
 } from './enter-password.form';
 import { useEnterPasswordStyles } from './enter-password.styles';
 
-const LOCK_TIME = 60_000;
-
-const checkTime = (i: number) => (i < 10 ? '0' + i : i);
-
-const getTimeLeft = (start: number, end: number) => {
-  const isPositiveTime = start + end - Date.now() < 0 ? 0 : start + end - Date.now();
-  const diff = isPositiveTime / 1000;
-  const seconds = Math.floor(diff % 60);
-  const minutes = Math.floor(diff / 60);
-
-  return `${checkTime(minutes)}:${checkTime(seconds)}`;
-};
-
 export const EnterPassword = () => {
   const styles = useEnterPasswordStyles();
 
-  const dispatch = useDispatch();
+  const { unlock, unlockWithBiometry } = useAppLock();
+
   const attempt = usePasswordAttempt();
   const timelock = usePasswordTimelock();
-  const { unlock, unlockWithBiometry } = useAppLock();
-  const lockLevel = LOCK_TIME * Math.floor(attempt / 3);
-  const [timeleft, setTimeleft] = useState(getTimeLeft(timelock, lockLevel));
-
   const { biometryType } = useBiometryAvailability();
   const handleResetDataButtonPress = useResetDataHandler();
-  const isDisabled = useMemo(() => Date.now() - timelock <= lockLevel, [timelock, lockLevel]);
+  const { isDisabled, timeleft } = usePasswordLock();
 
   const biometricsEnabled = useBiometricsEnabledSelector();
 
@@ -73,19 +56,6 @@ export const EnterPassword = () => {
     },
     [unlock, isDisabled, timeleft, timelock, attempt]
   );
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      if (Date.now() - timelock > lockLevel) {
-        dispatch(setPasswordTimelock.submit(0));
-      }
-      setTimeleft(getTimeLeft(timelock, lockLevel));
-    }, 1_000);
-
-    return () => {
-      clearInterval(interval);
-    };
-  }, [timelock, lockLevel]);
 
   useDelayedEffect(HIDE_SPLASH_SCREEN_TIMEOUT, () => void (isBiometryAvailable && unlockWithBiometry()), [
     isBiometryAvailable
