+++ src/screens/sync-account/after-sync-qr-scan/confirm-sync/confirm-sync.tsx
@@ -1,7 +1,6 @@
 import { Formik } from 'formik';
-import React, { FC, useEffect, useMemo, useState } from 'react';
+import React, { FC } from 'react';
 import { Text, View } from 'react-native';
-import { useDispatch } from 'react-redux';
 
 import { AttentionMessage } from '../../../../components/attention-message/attention-message';
 import { ButtonLargePrimary } from '../../../../components/button/button-large/button-large-primary/button-large-primary';
@@ -19,24 +18,10 @@ import { MaxPasswordAttemtps } from '../../../../config/system';
 import { FormBiometryCheckbox } from '../../../../form/form-biometry-checkbox/form-biometry-checkbox';
 import { FormCheckbox } from '../../../../form/form-checkbox';
 import { FormPasswordInput } from '../../../../form/form-password-input';
-import { setPasswordTimelock } from '../../../../store/security/security-actions';
-import { usePasswordAttempt, usePasswordTimelock } from '../../../../store/security/security-selectors';
+import { usePasswordLock } from '../../../../hooks/use-password-lock.hook';
 import { formatSize } from '../../../../styles/format-size';
 import { ConfirmSyncFormValues, ConfirmSyncValidationSchema } from './confirm-sync.form';
 import { useConfirmSyncStyles } from './confirm-sync.styles';
-
-const LOCK_TIME = 5_000;
-
-const checkTime = (i: number) => (i < 10 ? '0' + i : i);
-
-const getTimeLeft = (start: number, end: number) => {
-  const isPositiveTime = start + end - Date.now() < 0 ? 0 : start + end - Date.now();
-  const diff = isPositiveTime / 1000;
-  const seconds = Math.floor(diff % 60);
-  const minutes = Math.floor(diff / 60);
-
-  return `${checkTime(minutes)}:${checkTime(seconds)}`;
-};
 
 interface ConfirmSyncProps {
   onSubmit: (formValues: ConfirmSyncFormValues) => void;
@@ -45,12 +30,7 @@ interface ConfirmSyncProps {
 export const ConfirmSync: FC<ConfirmSyncProps> = ({ onSubmit }) => {
   const styles = useConfirmSyncStyles();
 
-  const dispatch = useDispatch();
-  const timelock = usePasswordTimelock();
-  const attempt = usePasswordAttempt();
-  const lockLevel = LOCK_TIME * Math.floor(attempt / 3);
-  const [timeleft, setTimeleft] = useState(getTimeLeft(timelock, lockLevel));
-  const isDisabled = useMemo(() => Date.now() - timelock <= lockLevel, [timelock, lockLevel]);
+  const { isDisabled, timeleft } = usePasswordLock();
 
   useNavigationSetOptions(
     {
@@ -59,19 +39,6 @@ export const ConfirmSync: FC<ConfirmSyncProps> = ({ onSubmit }) => {
     },
     []
   );
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      if (Date.now() - timelock > lockLevel) {
-        dispatch(setPasswordTimelock.submit(0));
-      }
-      setTimeleft(getTimeLeft(timelock, lockLevel));
-    }, 1_000);
-
-    return () => {
-      clearInterval(interval);
-    };
-  }, [timelock, lockLevel]);
 
   return (
     <Formik
